Module Alert_stdlib

Variants of Stdlib, but with signature items that can be restricted annotated with alerts.

There is one "variant" of the Stdlib interface implemented, namely Alerting_stdlib. Alerts are placed on signature items (types, values, and modules) so that those items can be restricted.

The following alerts are used:

physical_eq: all kinds of physical comparisons

debug_macro: macros like __LOC__

list_op: operations on lists

input_output: all kinds of real IO

impure: references, arrays, and other imperative programming constructs

unsafe: generally unsafe operations

By default, for a "safe" sandbox environment, the following alerts should be enabled: physical_eq, input_output, impure, unsafe. The remaining alerts are intended for specific situtations and exercises where more needs to be restricted.

module type Alerting_stdlib = sig ... end

Derived from Stdlib, but with items that can be restricted annotated with alerts.