Module Ast_check.Feature

type t =

| Array

| Mutable_member

| Object

| Loop

| Primitive

| Internal_name

| Alert_control

| Tail_mod_cons

module Set : Stdlib.Set.S with type elt = t

val to_identifier : t -> string

Unique identifier for this feature.

val of_identifier : string -> t

Provides the feature corresponding to a unique identifier returned by to_identifier.

val minimum : Set.t

Primitive and Internal_name, required to prevent circumventing restrictions.

val default : Set.t

Everything but Tail_mod_cons.

val all : Set.t

Everything.

val to_message : t -> string

Provide a message explaining why a feature is prohibited.