Module Common.Ctx_util

Context managers, loosely inspired by Python

type ('a, 'b, 'c) t = ('a -> ('b, exn) Stdlib.result) -> ('c, exn) Stdlib.result

A context manager of type ('a, 'b, 'c) t takes a continuation, and should feed the continuation a value of type 'a. Once the continuation returns with either a Ok of 'b or an exception, the continuation should perform cleanup, and may suppress the exception by producing a suitable result (of type 'c) instead. Typically, 'b = 'c.

This representation has the advantage that some existing functions library already implement this type (e.g. In_channel.with_open_text).

val with_context : ('a, 'b, 'c) t -> ('d -> 'e) -> 'f

with_context cm f runs f in the context manager cm

val let< : ('a, 'b, 'c) t -> ('a -> 'b) -> 'c

Let-syntax for with_context

val temp_file : ?temp_dir:string -> string -> string -> (string, 'a, 'b) t

As per Filename.temp_file. Removes the temporary file upon completion.

val openfile : string -> Unix.open_flag list -> Unix.file_perm -> (Unix.file_descr, 'a, 'b) t

As per Unix.openfile. Closes the file descriptor upon completion.

val set_signal : int -> Stdlib.Sys.signal_behavior -> (unit, 'a, 'b) t

As per Sys.set_signal. Restores the previous signal behavior on completion.

val sigprocmask : Unix.sigprocmask_command -> int list -> (unit, 'a, 'b) t

As per Unix.sigprocmask. Restores the previous mask on completion.

val timeout_unix : ?timer:Unix.interval_timer -> Mtime.span -> (unit, 'a, 'b option) t

val timed : (unit, 'a, 'b * Mtime.span) t

val capture_exceptions : ?filter:(exn -> bool) -> unit -> (unit, 'a, ('b, exn) Stdlib.result) t

val empty_context : 'a -> ('b -> 'c) -> ('d, 'e, 'f) t

val empty_context' : 'a -> ('b, 'c, 'd) t

val optional_context : some:('a -> ('b, 'c, 'd) t) -> 'e -> ('c -> 'd) -> 'f option -> ('b, 'c, 'd) t

val optional_timeout_unix : ?timeout:Mtime.span -> (unit, 'a, 'b option) t