Module Internal.Path_util

Path and file-related utility functions.

type walk_order =

| BottomUp

| TopDown

val walk : ?max_depth:int -> ?walk_order:walk_order -> ?follow_symlinks:bool -> string -> (string, [ `Access_file of string * string | `Unix_error of Unix.error * string * string ]) {C}1.Many.t

val (/) : string -> string -> string

val readdir_p : string -> string Stdlib.Seq.t

like Sys.readdir, but the returned entries are full paths

val is_code : ?condition:FileUtil.test_file -> FilePath.filename -> bool

Checks that either p is a regular file and condition holds for it, or p is a symlink to a regular file, and condition holds for the symlink.