Module Common.P_run

Even higher-level wrapper around Unix.open_process.

val wait_pid_timeout : Mtime.span -> int -> Unix.process_status option

Waits for the given process or times out.

type timeout_description = {

duration : Mtime.span;

signal : int;

kill_after : Mtime.span option;

}

Timeout after duration, sending signal. If kill_after is Some t, then send SIGKILL after waiting an additional t.

val timeout : ?signal:int -> ?kill_after:Mtime.span -> Mtime.span -> timeout_description

type phase =

| Completed

| TimedOut

| Killed
(*
Whether the process is running normally, has received the first signal, or has received a SIGKILL.
*)

type result = {

phase : phase;
(*
What phase was the process in when it completed?
*)

status : Unix.process_status;
(*
What status did the process exit with?
*)

stdout : string;

stderr : string;

elapsed : Mtime.span;

}

val result_is_ok : ?check_status:bool -> result -> bool

Did the subprocess exit normally? If check_status is false, ignore the exit code and check only for a timeout or signal

val pp_result : ?hide_stdout:bool -> ?hide_stderr:bool -> ?command_line:string list -> Stdlib.Format.formatter -> result -> unit

val build_env : (string * string) list -> string list

val p_run : ?timeout:timeout_description -> ?input:(Stdlib.Format.formatter -> unit) -> ?args:string list -> ?env:(string * string) list -> string -> result

Send the input to command with the given args. Note: uses temporary files provided by Filename.