Module Common.Path_util

Path and file-related utility functions.

val (/) : string -> string -> string

val readdir_p : string -> string Stdlib.Seq.t

like Sys.readdir, but the returned entries are full paths

val is_code : ?condition:FileUtil.test_file -> FilePath.filename -> bool

Checks that either p is a regular file and condition holds for it, or p is a symlink to a regular file, and condition holds for the symlink.

val mkdir : ?mode:Unix.file_perm -> ?exist_ok:bool -> string -> unit