Module Common.Util

Various utility functions.

val (%) : ('a -> 'b) -> ('c -> 'a) -> 'c -> 'b

Function composition, f % g is f \circ g.

val (%>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

Reverse function composition, f %> g is g \circ f.

val peek : ('a -> unit) -> 'a -> 'a

peek f x is (f x; x).

val singleton : 'a -> 'a list

One-element list.

val null : 'a list -> bool

val uncons : 'a list -> ('a * 'a list) option

val unsnoc : 'a list -> ('a list * 'a) option

val list_equal : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

Same as List.equal, but allows lists of different types

val update_assoc : ('a option -> 'a option) -> 'b -> ('b * 'a) list -> ('b * 'a) list

val or_option : 'a option -> 'a option -> 'a option

val filter_option : ('a -> bool) -> 'a option -> 'a option

val if_option : bool -> 'a -> 'a option

val string_contains : needle:string -> string -> bool

string_contains ~needle haystack: does needle exist as a substring of haystack? Naive \mathcal{O}(nm) implementation.

val (~$) : 'a -> ('a -> 'b) -> 'b

(~$ x) f is f x, e.g.: List.map (~$ 2) [List.nth ["a"; "b"; "c"]; String.sub "def" 1] returns ["c"; "ef"]

type exn_info = {

exn : exn;

backtrace : Stdlib.Printexc.raw_backtrace;

}

Exception and its associated (raw) backtrace.

val raise_exn_info : exn_info -> 'a

val try_to_result : ('a -> 'b) -> 'a -> ('b, exn_info) Stdlib.result

val unresult_exn : ('a, exn_info) Stdlib.result -> 'a

val run_main : (string -> string list -> int) -> 'a

Timeouts

val min_span : Mtime.span -> Mtime.span -> Mtime.span

exception Timeout

val span_to_float_s : Mtime.span -> float

Can overflow for large t

val timeout_unix : ?timer:Unix.interval_timer -> Mtime.span -> ('a -> 'b) -> 'a -> 'b option

timeout_unix t f is Some (f ()) if the call to f terminates within t, otherwise None. The timeout is triggered by a Unix signal.

Notes:

- Does not interrupt f until an allocation occurs, which may be never.

- Any existing Unix timer is suspended for the duration of the call to f.

+ If f is interrupted, it will not be left running in the background.

val timed : ('a -> 'b) -> 'a -> 'b * Mtime.span