Module Signature_builder

Library and PPX rewriter for building interfaces from existing interfaces by selectively adding and removing interface items.

See Parse for the full grammar for inclusion specifications.

Loading saved signatures

type raw_signature = [

| `IntfMarshal of string
(*
Marshalled Ppxlib.interface
*)

| `IntfSrc of string
(*
Plain source code of an interface
*)

]

type raw_signature_info = (string list * raw_signature) list

type signature_infos = (string * signature_info) list

and signature_info =

| Node of Ppxlib.signature option * signature_infos

val infos_add_signature : string -> string list -> Ppxlib.signature -> signature_infos -> signature_infos

val info_add_signature : string list -> Ppxlib.signature -> signature_info -> signature_info

val infos_get_signature : string -> string list -> signature_infos -> Ppxlib.signature option

val info_get_signature : string list -> signature_info -> Ppxlib.signature option

val load_raw_signature : raw_signature -> Ppxlib.signature

val load_signature_info : raw_signature_info -> signature_info

Signature of Stdlib and Thread/Event

val stdlib_signature_info : signature_info

val threads_signature_info : signature_info

PPX rewriter

module Ordered_set : sig ... end

type kind =

| Value

| Type

| Exception

| Module

| ModuleType

| Class

| ClassType

type signature_item_pat = {

names : Ppxlib.longident_loc list;

kind : kind option;

loc : Ppxlib.location;

}

type include_specs = include_spec list

and include_spec =

| Include_from of {

from : Ppxlib.longident_loc;

items : signature_item_pat Ordered_set.t;

loc : Ppxlib.Location.t;

}

| Include_group of {

attributes : Ppxlib.attribute list;

items : include_specs;

}

module Parse : sig ... end

Parser for include specifications.

val is_signature_item_selectable : Ppxlib.signature_item -> bool

type signature_item_name =

| Unnamed

| Lident of Ppxlib.longident

| String of string

val match_lident_name : Ppxlib.longident -> signature_item_name -> bool

Does the given longident (as a search pattern) match the given name?

val signature_item_names : Ppxlib.signature_item -> signature_item_name list

val match_kind : kind -> Ppxlib.signature_item -> bool

val match_signature_item_pat : signature_item_pat -> Ppxlib.signature_item -> bool

val eval_pat_ordered_set : Ppxlib.location -> Ppxlib.signature_item list -> signature_item_pat Ordered_set.t -> Ppxlib.signature_item list

val modify_attributes : (Ppxlib.attributes -> Ppxlib.attributes) -> Ppxlib.signature_item -> Ppxlib.signature_item

val eval_include_specs : signature_infos -> include_specs -> Ppxlib.signature_item list

val eval_include_spec : signature_infos -> include_spec -> Ppxlib.signature_item list

val default_sig_infos : (string * signature_info) list

val ppx_include : ?sig_infos:signature_infos -> ctxt:Ppxlib.Expansion_context.Extension.t -> Ppxlib.expression -> Ppxlib__.Import.signature_item

PPX intended for use an a Ppxlib extender. The default sig_infos argument contains descriptions for the standard library under "stdlib" and for threads/events under "threads".