Module Signature_builder.Parse

Parser for include specifications.

To use the signature builder, first enable the PPX rewriter. With dune, add the PPX rewriter to your component's dune file:

(preprocess (pps less-power.signature-builder-ppx))

Then, within a signature, use an extension point to include signature items, for example:

module type Foo : sig
  [%%include stdlib.stdlib (fst, snd)]
end

Inclusion specifications can be more complex. The full grammar is as follows:

include-specs := include-spec | '()' | '(' include-spec {',' include-spec} ')'
include-spec :=
  | from-spec items-spec
  | '{' 'items =' include-specs ';' ['attributes = __' {attribute} ';'] '}'
from-spec := (* qualified path to signature-info to take items from *)
  lowercase-ident {'.' lowercase-ident}+
items-spec := (* ordered set of items *)
  item-spec | '!standard' | items-spec '-' items-spec | '()' | '(' items-spec {',' items-spec} ')'
item-spec := item-names ['@' item-options]
item-names := item-name | '[' {item-name ';'}+ ']'
item-name := (* refer to values, types, modules, ... *)
  lowercase-ident | capitalized-ident | '[]'

The following more complex example includes t and to_string_default from the standard Printexc module, then also includes everything except that type and that value, with each item given the unsafe attribute:

[%%include
  stdlib.printexc (t, to_string_default),
  { attributes = __ [@alert unsafe "This item is not permitted"];
    items = stdlib.printexc (!standard - (t, to_string_default)) }
]

To disambiguate signature items of different kinds with the same name, use the syntax identifier @ { kind = "kind"}, where kind is one of value, type, exception, module, module-type. For example, to include the standard ref function:

[%%include stdlib.stdlib (ref @ { kind = "value" })]

The preprocessor will refuse to include types which don't include a type equation (e.g. abstract types and types defined only by constructors). This is because this will create copies of the type, which is generally not the intention. In this case, types must be defined manually without the inclusion mechanism, e.g.:

type ('a, 'b) result = ('a, 'b) Stdlib.result = Ok of 'a | Error of 'b

val ordered_set : Ppxlib_ast.Ast.expression -> Ppxlib_ast.Ast.expression Ordered_set.t

val list : Ppxlib.expression -> Ppxlib_ast.Ast.expression list

val kind : Ppxlib.expression -> kind

val signature_item_pat : Ppxlib_ast.Ast.expression -> signature_item_pat

val dotted_name : Ppxlib.expression -> Ppxlib.longident

val include_specs : Ppxlib.expression -> include_specs

val include_spec : allow_comma:bool -> Ppxlib.expression -> include_spec