Module Std_overrides

Ready-to-use modules for replacing the default pervasives, using -open.

module Hide_stdlib_variants : sig ... end

Prevent access to the full variant library. Intended to be opened at the top level to hide the Stdlib_variants module, and a couple others.

module Stdlib_safe : sig ... end

Puts the safe parts of Stdlib into scope, from Stdlib_components.Stdlib_safe

module Stdlib_alerting : sig ... end

Puts an alerting Stdlib into scope, from Stdlib_alerts.Stdlib_alerting.