Module Std_overrides.Hide_stdlib_variants

Prevent access to the full variant library. Intended to be opened at the top level to hide the Stdlib_variants module, and a couple others.

This is included in the other overrides modules, so you usually don't need to pass it to the -open flag yourself. However, you should use it if you define your own override module based on less-power stdlib variants.

These modules are intentionally empty.

module Stdlib_variants : sig ... end

module Stdlib_components : sig ... end

module Stdlib_alerts : sig ... end

module Threads_alerts : sig ... end