Module Stdlib_alerts

Variants of Stdlib, but with signature items that can be restricted annotated with alerts.

There is one "variant" of the Stdlib interface implemented, namely Stdlib_alerting. Alerts are placed on signature items (types, values, and modules) so that those items can be restricted.

The following alerts are used:

physical_eq: all kinds of physical comparisons

debug_macro: macros like __LOC__

list_op: operations on lists

input_output: all kinds of real IO

impure: references, arrays, and other imperative programming constructs

unsafe: generally unsafe operations

By default, for a "safe" sandbox environment, the following alerts should be enabled: physical_eq, input_output, impure, unsafe. The remaining alerts are intended for specific situtations and exercises where more needs to be restricted.

Similar modules for Thread and Event are in Threads_alerts to avoid unnecessary dependencies.

module type Char_alerting = sig ... end

module type Digest_alerting = sig ... end

module type Filename_alerting = sig ... end

module type Float_alerting = sig ... end

module type Hashtbl_alerting = sig ... end

module type List_alerting = sig ... end

module type ListLabels_alerting = sig ... end

module type MoreLabels_alerting = sig ... end

module type Printexc_alerting = sig ... end

module type Printf_alerting = sig ... end

module type Scanf_alerting = sig ... end

module type Seq_alerting = sig ... end

module type StringLabels_alerting = sig ... end

module type StdLabels_alerting = sig ... end

module type String_alerting = sig ... end

module type Uchar_alerting = sig ... end

module type Stdlib_alerting = sig ... end

Derived from Stdlib, but with items that can be restricted annotated with alerts.

module Stdlib_alerting : Stdlib_alerting

Stdlib, but with the above module type.