Module Stdlib_components.SafeAliases

SAFE Safe modules from the standard library, and unsafe modules restricted to their safe functions.

module Bool = Stdlib.Bool

module Char = Stdlib.Char

module Complex = Stdlib.Complex

module Either = Stdlib.Either

module Float : sig ... end

As module type of Float, but without Array and ArrayLabels.

module Fun = Stdlib.Fun

module Int = Stdlib.Int

module Int32 = Stdlib.Int32

module Int64 = Stdlib.Int64

module Lazy = Stdlib.Lazy

module List = Stdlib.List

module ListLabels = Stdlib.ListLabels

module Map = Stdlib.Map

module MoreLabels : sig ... end

module Nativeint = Stdlib.Nativeint

module Option = Stdlib.Option

module Printexc : sig ... end

module Printf : sig ... end

module Result = Stdlib.Result

module Scanf : sig ... end

module Seq : sig ... end

As module type of Seq, but without once, Forced_twice and to_dispenser.

module Set = Stdlib.Set

module StdLabels : sig ... end

module String : sig ... end

As module type of String, but without some unsafe buffer-related functions.

module StringLabels : sig ... end

module Uchar = Stdlib.Uchar

module Unit = Stdlib.Unit