Module Stdlib_variants

Variants of Stdlib with restricted or modified functionality.

Stdlib_components Adapted from Stdlib, with different components in different modules. In particular, a safe (pure) subset of OCaml features can be enabled by importing a subset of modules defined here.

module Stdlib_components = Stdlib_components

Alert_stdlib Like Stdlib, but with items that can be restricted annotated with alerts.

module Alert_stdlib = Alert_stdlib

module Hide_stdlib_variants : sig ... end

module SafeStdlib : sig ... end

SAFE Everything safe from the standard library, including everything safe from sub-modules of Stdlib.