Module Stdlib_variants.Stdlib_components

Adapted from Stdlib, with different components in different modules. In particular, a safe (pure) subset of OCaml features can be enabled by importing a subset of modules defined here.

module Exceptions : sig ... end

SAFE raise primitives and standard exceptions

module Composition : sig ... end

SAFE Composition operators

module Debugging : sig ... end

SAFE Debugging macros

module Comparisons : sig ... end

SAFE Polymorphic comparisons. Does not include the impure (==) and (!=), which are in PhysicalComparisons.

module PhysicalComparisons : sig ... end

UNSAFE The impure (==) and (!=)

module BooleanOperations : sig ... end

SAFE Boolean operations

module IntegerOperations : sig ... end

SAFE Integer operations

module FloatingPointOperations : sig ... end

SAFE Floating-point operations

module StringOperations : sig ... end

SAFE String operations -- more in Stdlib module String

module CharOperations : sig ... end

SAFE Character operations -- more in module Char

module UnitOperations : sig ... end

SAFE Unit operations

module PairOperations : sig ... end

SAFE Pair operations

module References : sig ... end

UNSAFE References

module Result : sig ... end

SAFE Result type

module StringConversion : sig ... end

SAFE String conversion functions

module ListOperations : sig ... end

SAFE List operations -- more in module List

module IOOperations : sig ... end

UNSAFE I/O operations. This is regular, unmocked, IO.

module Formats : sig ... end

SAFE Formats

module Exit : sig ... end

UNSAFE Functions exiting the program at_exit, exit

module SafeAliases : sig ... end

SAFE Safe modules from the standard library, and unsafe modules restricted to their safe functions.