Module Test_lib.Grading

Provide a grade from a list of passed/failed tests and a grading scheme.

type 'a clamp = {

min : 'a;

max : 'a;

}

A range of values, inclusive.

val clamp : 'a clamp -> 'a -> 'a

clamp { min = x; max = y } z constrains z between x and y.

val clamp_opt : 'a clamp option -> 'a -> 'a

clamp None x is x, otherwise as clamp.

type grading_result = {

text : string;

points : int;

max_points : int;

}

The result of grading a list of tests.

type tests = (string * bool) list

Pairs of test_name, passed.

type grading_criterion =

| Passed of string
(*
Did the given test pass?
*)

| Failed of string
(*
Did the given test fail?
*)

| Not of grading_criterion
(*
Negation (logical not).
*)

| OneOf of grading_criterion list
(*
Disjunction (logical or).
*)

| AllOf of grading_criterion list
(*
Conjunction (logical and).
*)

| Constant of bool
(*
Logical constant.
*)

Logical formulas over test results.

val implies : grading_criterion -> grading_criterion -> grading_criterion

Logical implication: implies a c = OneOf [Not a; c].

val string_of_grading_criterion : ?indent:int -> grading_criterion -> string

type grading =

| Points of {

title : string;

points : int;

test_case : grading_criterion;

reason : tests -> grading_criterion -> string;

}
(*
If the criterion test_case evaluates to true, assign a grade of points, otherwise 0.
*)

| Group of {

title : string;

items : grading list;

max_points : int clamp option;

skip : string option;
(*
If skip = Some r, skip this group (assign a grade of 0) with the reason r.
*)

}
(*
A group of grading schemes. Assign the sum of grades from the sub-items, with an optional minimum/maximum number of points.
*)

| Conditional of {

condition : grading_criterion;

message : string;

content : grading;

}
(*
Assign a grade according to content if the condition evaluates to true, otherwise 0.
*)

Grading scheme. Defines how to assign a numeric grade for a given set of test results.

exception No_reason

val points : ?reason:(tests -> grading_criterion -> string option) -> ?penalty:bool -> string -> int -> grading_criterion -> grading

Wrapper for Points. If reason returns None (the default), the displayed reason is decided by penalty and points:

"PASS"/"FAIL" if points is non-negative, or penalty is true

If points is negative, or penalty is false, then the grading criterion must evaluate to true and the reason is "PENALTY", otherwise No_reason is raised during evaluation.

If reason returns a value, that is used directly instead.

val assertion : ?message:string -> ?title:string -> int -> grading_criterion -> grading

Wrapper for Points, intended for debugging tests. If the given grading_criterion does not evaluate to true, the message is given as a reason along with a print-out of the criterion. Useful for verifying invariants between tests, e.g. that passing a more comprehensive test implies passing a test that tests a subset of functionality.

val points_p : ?reason:(tests -> grading_criterion -> string option) -> ?penalty:bool -> string -> int -> string -> grading

val points_f : ?reason:(tests -> grading_criterion -> string option) -> ?penalty:bool -> string -> int -> string -> grading

val points_c : ?reason:(tests -> grading_criterion -> string option) -> ?penalty:bool -> string -> int -> bool -> grading

As per points, but with the grading criterion set to Passed, Failed, or Constant, respectively.

val conditional : grading_criterion -> string -> grading -> grading

val conditional_c : bool -> string -> grading -> grading

Wrappers for Conditional, with the given condition or a constant.

val group : ?skip:string -> ?max_points:int clamp -> string -> grading list -> grading

Wrapper for Group.

val prettify_results_basic : string -> unit

val prettify_results : ?grading:grading -> ?points_step_count:int -> string -> unit

Main entry point: prettify_results ~grading file_name reads the JUnit XML file in file_name, evalutes the given grading scheme and writes the result back to the same file.