Module Test_lib.OUnit_util

Utilities for converting from QCheck to OUnit, and other OUnit-related functionality.

val default_qcheck_timeout : Mtime.span

val default_rand : Stdlib.Random.State.t

val of_qcheck : ?timeout:Mtime.span -> QCheck2.Test.t -> OUnitTest.test

Convert a QCheck test to OUnit, overwriting the timeout. Automatically prevents excessive shrinking from timing out tests, by interrupting shrinking and showing the smallest value found so far once the timeout expires.

val map_test : ?name:(string -> string) -> ?test_fun:(OUnitTest.test_fun -> OUnitTest.test_fun) -> ?test_length:(OUnitTest.test_length -> OUnitTest.test_length) -> OUnitTest.test -> OUnitTest.test

type visibility =

| PassFail
(*
Test is always run, but the detailed result is not shown. Instead, a failure is indicated with a generic message. Also known as "secret" tests.
*)

| None
(*
Test is not run. Instead, the test fails and a generic message indicates that the test was not run. Also known as "hidden" tests.
*)

Visibility of tests after hiding.

val message_of_visibility : visibility -> string

val hide_test : ?visibility:visibility -> OUnitTest.test -> OUnitTest.test

Hide a test according to the visibility argument.

val conditional_test : (string * bool) list -> OUnitTest.test -> OUnitTest.test

val conditional_test1 : string -> bool -> OUnitTest.test -> OUnitTest.test

val join_tests : OUnitTest.test -> OUnitTest.test

Join a test tree into a single test which only passes if all individual tests in the tree pass. Students will only be able to see the result of the first test that failed. Timeouts of individual tests are summed to produce a single timeout, which is applied to the combined test. This changes the timeout behavior of the tests. Intended for grouping tests, when grading is performed solely by assigning points to passed tests.

For more complex grouping, use the Grading module.