Module Test_lib.QCheck_util

Helpers for QCheck-based tests.

val default_computation_timeout : Mtime.span

val is_user_fail : exn -> string option

Workaround to check if an exception is QCheck2.Test.User_fail, since that exception is not exposed.

val assert_equal : ?context:string -> ?timeout:Mtime.span -> eq:('a -> 'b -> 'c option) -> printers:(('d -> string) * ('e -> string)) -> (unit -> 'f) -> (unit -> 'g) -> ('h, string) Stdlib.result

Check if values are equal. Report a failure message as Error if not. This monadic form is useful for properties that check nested functions, i.e. calling submission code contains functions that should once again be called.

Note: with timeout, uses Common.Util.timeout_unix, and thus nesting (within expected_f or actual_f) will not work as expected.

parameter context
Add this string as context before failure message

parameter eq
Check for equality, return Some x if equal or None if not

returns
The result of eq, if not None, as Ok.

val lift_eq : ('a -> 'b -> bool) -> 'c -> 'd -> unit option

Turn an equality function into the form required for assert_equal.

val assert_equal' : ?eq:('a -> 'b -> unit option) -> printer:('b -> string) -> ?context:string -> ?timeout:Mtime.span -> (unit -> 'a) -> (unit -> 'b) -> (unit, string) Stdlib.result

Like assert_equal, but when the types to compare are the same.

val fail_reportf_notrace : ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'c

val fail_report_notrace : string -> 'a

val report_result : ('a, string) Stdlib.result -> bool

After performing one or more comparisons with assert_equal, report a failure as a QCheck failure, or just return true otherwise.

val make_test_singles : ?name:string -> ?print:'a QCheck2.Print.t -> 'a0 list -> ('a1 -> bool) -> QCheck2.Test.t

Check a property against multiple, fixed inputs.

val make_test_single : ?name:string -> ?print:'a QCheck2.Print.t -> 'b -> ('a -> bool) -> QCheck2.Test.t

Check a property against a fixed input.