Module Test_runner

Library for creating a runner, which then drives the testing process.

module Task : sig ... end

Set up, build, and run tests using first-class tasks.

module Std_task : sig ... end

Common tasks and tasks for the standard test setup.

module Entry_point : sig ... end

Entry points for creating test driver executables from a task tree.

module P_run = Common.P_run

Re-export of Common.P_run. Tasks spawning subprocesses can use parts of this module.