Module Test_runner.Entry_point

Entry points for creating test driver executables from a task tree.

val build_root : string Cmdliner.Term.t

val safe : bool Cmdliner.Term.t

val mtime_s : Mtime.span Cmdliner.Arg.conv

val mk_timeout : names:string list -> task:string -> default:Mtime.span -> Mtime.span Cmdliner.Term.t

val build_timeout : Mtime.span Cmdliner.Term.t

val probe_timeout : Mtime.span Cmdliner.Term.t

val test_timeout : Mtime.span Cmdliner.Term.t

val mk_timestamp : names:string list -> ?doc:string -> ?absent:string -> (Cmdliner.Arg.info -> 'a Cmdliner.Arg.t) -> 'a Cmdliner.Term.t

val timestamp_now : float option Cmdliner.Term.t

val mk_timestamp_start_end : when_:string -> names:string list -> default:float -> float Cmdliner.Term.t

val exercise_start : float Cmdliner.Term.t

val exercise_end : float Cmdliner.Term.t

val term_of_runner : (Std_task.cfg -> 'a) -> 'a Cmdliner.Term.t

val build_report_junit : ?time:Mtime.span -> string option -> Junit.t

Testcase.error if message is Some _, otherwise Testcase.pass

val task_runner : ?build_report_path:string -> ?report_success_details:bool -> (Std_task.cfg -> (unit, unit) Task.tree) -> Std_task.cfg -> unit

Creates a task from the config, and runs it. Reports the result as a test case (in build_report_path, default test-reports/build.xml). If report_success_details is false (default), no detailed build report is shown when the build succeeds. Otherwise, the detailed report is always shown, which requires that the test case is marked as a failure.

val command_of_term : 'a Cmdliner.Term.t -> 'a Cmdliner.Cmd.t

val run_task_main : ?build_report_path:string -> ?report_success_details:bool -> ?exit:(Cmdliner.Cmd.Exit.code -> 'a) -> (Std_task.cfg -> (unit, unit) Task.tree) -> 'a