Module Test_runner.Std_task

Common tasks and tasks for the standard test setup.

val f : ('a, unit, string) Stdlib.format -> 'a

Common tasks

val contains_symlink : string -> bool

Check if the directory contains a symlink. The directory itself may be a symlink.

module Messages : sig ... end

type cfg = {

build_root : FilePath.filename;
(*
relative paths are OK
*)

safe : bool;

build_timeout : Mtime.span;

test_timeout : Mtime.span;

probe_timeout : Mtime.span;

timestamp_now : float option;
(*
Override current time, UTC
*)

exercise_start : float;
(*
UTC time
*)

exercise_end : float;
(*
UTC time
*)

}

val check_contains_symlink : cfg -> string -> ('a, unit) Task.tree

val write_file_pp : ?label:string -> cfg -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> ('b, unit) Task.tree

val write_file_f : ?label:string -> cfg -> string -> ('a, Stdlib.Format.formatter, unit, ('b, unit) Task.tree) Stdlib.format4 -> 'a

val write_file_str : ?label:string -> cfg -> string -> string -> ('a, unit) Task.tree

val configure_show_hidden : cfg -> string -> ('a, unit) Task.tree

val remove_files : ?condition:FileUtil.test_file -> cfg -> string -> ('a, unit) Task.tree

Task: remove the files which match condition (default: .ml files). Not recursive: only files in p will be deleted, not in subdirectories.

val copy_files : ?condition:FileUtil.test_file -> cfg -> string -> string -> ('a, unit) Task.tree

Task: copy the files which match condition (default: .ml files). Not recursive: only files in src0 will be copied, not in subdirectories.

val load_files : cfg -> ?condition:FileUtil.test_file -> string -> string -> ('a, unit) Task.tree

Composition of remove_files and load_files.

val make_clean_directory : ?remove:FileUtil.test_file -> cfg -> string -> ('a, unit) Task.tree

val make_test_report_directory : cfg -> string -> (unit, unit) Task.tree

val checker : cfg -> ?prohibited:Ast_check.Feature.Set.t -> ?limit:int -> ?check1:FileUtil.test_file -> ?check:FileUtil.test_file -> string -> ('a, unit) Task.tree

Run the AST-checker as a task.

val subprocess_run : ?timeout:Common.P_run.timeout_description -> ?args:string list -> ?env:(string * string) list -> string -> ('a, Common.P_run.result) Task.tree

Run a process as a task

val subprocess_result : ?hide_stdout:bool -> ?hide_stderr:bool -> ?command_line:string list -> ?error_message:string -> ?check_status:bool -> ?dump_output:bool -> unit -> (Common.P_run.result, Common.P_run.result) Task.tree

Handle the result of a subprocess task.

type subprocess_options = {

timeout : Common.P_run.timeout_description option;

hide_stdout : bool;

hide_stderr : bool;

error_message : string option;

check_status : bool;

dump_output : bool;

}

val subprocess_options : ?timeout:Common.P_run.timeout_description -> ?hide_stdout:bool -> ?hide_stderr:bool -> ?error_message:string -> ?check_status:bool -> ?dump_output:bool -> unit -> subprocess_options

val subprocess : cfg -> ?options:subprocess_options -> ?args:string list -> ?env:(string * string) list -> string -> ('a, Common.P_run.result) Task.tree

val dune : cfg -> ?options:subprocess_options -> root:string -> ?args:string list -> ?env:(string * string) list -> string -> ('a, Common.P_run.result) Task.tree

val timeout_for : cfg -> [< `Build | `Probe | `Test ] -> Mtime.span

Standard tasks

These are high-level tasks built upon the primitives defined above. They work with the recommended layout of tests, submission and sample solution. However, the low-level tasks defined above can also be used as needed.

val std_test_report_dir : string

val std_setup : cfg -> (unit, unit) Task.tree

Set up the test environment.

val std_check : cfg -> ('a, unit) Task.tree

Run the limitation checker.

val std_build1 : cfg -> ?submission:bool -> ?env:(string * string) list -> string -> ('a, unit) Task.tree

Build something in the tests/ directory with dune. If submission is false (default), error output is hidden.

val std_exec1 : cfg -> ?phase:[< `Build | `Probe | `Test ] -> ?args:string list -> ?env:(string * string) list -> string -> ('a, unit) Task.tree

Run something in the tests/ directory with dune exec.

This task passes the --no-build flag to dune. As such:

The executable must be built before running this task (e.g. with std_build1).

There is no risk of showing the build output to the student.

If phase is given, sets a timeout.

val std_build : cfg -> (unit, unit) Task.tree

Build the assignment library and test binary.

val std_exec_test : cfg -> ?env:(string * string) list -> output_junit_file:string -> string -> (unit, unit) Task.tree

Run a test executable in the standard setup. See the note about building first in std_exec1. Assumes the test executable is configured to exit immediately if no command line arguments are passed. We use this to check that the top level code of the submission does not contain any long-running code. If it does, there's no point in running the tests, since the significantly longer test timeout will just kill them. We don't want to prohibit top-level code entirely, since it would prevent things like let impl0 = impl 0.

Thus, first runs the given executable without any arguments, then runs it a second time, passing the given path as the output JUnit file.

val std_test : cfg -> (unit, unit) Task.tree

As std_exec_test, fixed to test/test.exe with output to test-reports/results.xml.

Probe-related tasks (partial type correctness checking)

type probe_result = {

name : string;
(*
the name given to the probe definition
*)

ok : bool;
(*
did the check pass?
*)

message : string;
(*
compiler output
*)

}

val probe_result_of_prun_result : name:string -> Common.P_run.result -> probe_result

val defines_of_probe_result : probe_result -> string list

val is_probe_name : Stdlib.String.t -> bool

[A-Z_]*

val load_probe_items : cfg -> defs:string -> ('a, (string * Parsetree.structure) list) Task.tree

parse the probe items from the given path

val probe_dune : name:string -> dep:string -> Stdlib.Format.formatter -> unit

val probe_dir_path : string -> name:string -> string

val gen_probe : cfg -> data_dir:string -> dep:string -> name:string -> Parsetree.structure -> (unit, unit) Task.tree

val run_probe : cfg -> root:string -> data_dir:string -> name:string -> ('a, probe_result) Task.tree

unlike the other tasks, data_dir needs to be relative to root here

val handle_probe_results : cfg -> args_file:string -> (probe_result list, unit) Task.tree

val probe : cfg -> root:string -> defs:string -> dep:string -> data_dir:string -> args_file:string -> ('a, unit) Task.tree

all paths are relative to root

val std_probe : cfg -> ('a, unit) Task.tree