Module Test_runner.Task

Set up, build, and run tests using first-class tasks.

Tasks define actions to run when testing submissions. Examples include building student code, checking for forbidden syntax elements, or running the generated test executable.

This module provides tasks as trees (type Tree.t), so tasks can be more easily re-used and composed.

Running tasks automatically handles exceptions and timeouts.

Task trees and task lists

type ('a, 'b) tree = {

label : string option;

timeout : Mtime.span option;

task : 'a -> 'b k;

}

A single task, possibly with subtasks (see k).

and _ k =

| Cont : 'a * ('a, 'b) tasks -> 'b k

Task continuation of type 'b k: Upon completion, a task produces a value of some type 'a, and a list of subtasks from 'a to 'b.

and (_, _) tasks =

| Nil : ('a, 'a) tasks

| Cons : ('a, 'b) tree * ('b, 'c) tasks -> ('a, 'c) tasks

A list of tasks, which start by accepting a value of type 'a and produce a result of type 'b.

val tree : ?label:string -> ?timeout:Mtime.span -> ('a -> 'b k) -> ('c, 'd) tree

val nil : ('a, 'a) tasks

val (@:>) : ('a, 'b) tree -> ('b, 'c) tasks -> ('a, 'c) tasks

val cons : ('a, 'b) tree -> ('b, 'c) tasks -> ('a, 'c) tasks

val app : 'a 'b 'c. ('a, 'b) tasks -> ('b, 'c) tasks -> ('a, 'c) tasks

Join two lists of tasks.

val of_list : ('a, 'b) tree list -> ('b, 'c) tasks

Create a task list from a regular list. All tasks must have the same input and output type (e.g. unit); if that is not desired, use the tasks constructors directly.

val singleton : ('a, 'b) tree -> ('a, 'c) tasks

val snoc : ('a, 'b) tasks -> ('c, 'd) tree -> ('a, 'e) tasks

val then_ : ('a, 'b) tree -> ('c, 'd) tree -> ('c, 'e) tasks

then_ t2 t1 is the task list [t1; t2]

val with_ : ?timeout:Mtime.span -> ?label:string -> ('a, 'b) tree -> ('c, 'd) tree

type (_, _) unsnoc =

| Unsnoc : ('a, 'b) tasks * ('b, 'c) tree -> ('a, 'c) unsnoc
(*
Unsnoc ts lst is the list snoc ts lst, but split into ts and lst. The list ts contains all elements of the list except the last one, and lst is the last element.
*)

val unsnoc : 'a 'b 'c. ('a, 'b) tree -> ('b, 'c) tasks -> ('a, 'c) unsnoc

unsnoc t1 ts splits the non-empty list cons t1 ts into its initial elements and final element, see unsnoc.

Task runner

Task failures

exception Task_failure of Stdlib.Format.formatter -> unit

type task_failure =

| Timed_out

| Exception_raised of exn

| Task_failed of Stdlib.Format.formatter -> unit

val pp_task_failure : Stdlib.Format.formatter -> task_failure -> unit

val failf : ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'c

Calling failf "..." ... will raise an exception that aborts the task runner with a formatted message.

val fail : string -> 'a

Like failf, but with a plain string.

Task results

type summary = {

label : string option;

elapsed_task : Mtime.span;
(*
does not include subtasks
*)

elapsed_total : Mtime.span;
(*
includes subtasks
*)

subtasks : summary list;

}

Runner implementation

type 'a st_in = {

x : 'a;

timeout : Mtime.span option;

}

type 'a st_out = ('a, task_failure) Stdlib.result

If the last task completed successfully with a value of x, then Ok x. Otherwise the failure reason of the task that failed.

val update_state_in : 'a -> Mtime.span -> 'b st_in -> 'c st_in

val run_tasks : 'a 'b. 'a st_in -> ('a, 'b) tasks -> 'b st_out * summary list

Run a list of tasks.

val run_tree : 'a 'b. 'a st_in -> ('a, 'b) tree -> 'b st_out * summary

Run a tree of tasks.

val run : ?timeout:Mtime.span -> ('a, 'b) tree -> 'c -> 'b st_out * summary

run ~timeout t x runs the task tree t with the initial input x, and the given timeout (default: no timeout).

Task helpers

val task1 : ?timeout:Mtime.span -> ?label:string -> ('a -> 'b) -> ('c, 'd) tree

Run a single function as a task, without any subtasks

val group : ?timeout:Mtime.span -> ?label:string -> ('a, 'b) tasks -> ('c, 'd) tree

Create a task tree from a list of tasks, the task itself will do nothing but return the list of subtasks.

val then_map_k : 'a 'b. ('a -> 'b) -> 'a k -> 'b k

val then_map : 'a 'b 'c. ('b -> 'c) -> ('a, 'b) tree -> ('a, 'c) tree

Modify the final result of a task tree, without creating an additional task. Note: since f runs inside the final task, its side-effects should be minimal: any exceptions will be raised inside an existing task, and the existing task's timeout will include f. If that is undesirable, run f inside a new subtask with then_subtask.

val then_subtask : ('a, 'b) tree -> ('c, 'd) tree -> ('e, 'f) tree

val ignore : ('a, 'b) tree -> ('a, unit) tree

Ignore the result of a task (replace it with ()).

val first_map : ('a -> 'b) -> ('c, 'd) tree -> ('e, 'f) tree

val first_const : 'a -> ('b, 'c) tree -> ('d, 'c) tree

val first_ignore : (unit, 'a) tree -> ('b, 'a) tree

val pure : 'a -> ('b, 'c) tree

val accumulator : ('a -> 'b -> 'c) -> (unit, 'd) tree -> ('e, 'f) tree

val spawning : ('a -> (unit, 'b) tasks) -> ('c, 'd) tree

spawning k runs as the list of tasks produced by applying k to the task's input

val accumulating : ('a -> 'b -> 'a) -> (unit, 'c) tree list -> ('a0, 'a1) tree list

accumulating f ts uses f to combine the output of each task into a running accumulator value

val collecting : (unit, 'a) tree list -> ('b list, 'b list) tree list

collecting ts pushes the outputs of each task onto a list

val collecting' : (unit, 'a) tree list -> ('b, 'c list) tasks

collecting' is like collecting, except with an empty initial list and a final List.rev step so that the results are in order; the result is a tasks rather than a list

val pp_summary : failure:bool -> ?show_anon:bool -> unit -> summary Fmt.t

val pp_state_out : 'a Fmt.t -> Stdlib.Format.formatter -> 'b st_out -> unit