Module Test_runner.Task_tree

Task trees represent a series of tasks to be run.

exception Task_failure of Stdlib.Format.formatter -> unit

val failf : ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'c

val fail : string -> 'a

type task_failure =

| Timed_out of Mtime.span

| Exception_raised of exn

| Task_failed of Stdlib.Format.formatter -> unit

val pp_task_failure : Stdlib.Format.formatter -> task_failure -> unit

type ctx = {

label : string option;

}

type ctxs = ctx list

type ('a, 'b) l =

| Nil : ('a, 'a) l

| Cons : ('a, 'b) t * ('b, 'c) l -> ('a, 'c) l

A list of tasks, which start by accepting a value of type 'a and produce a result of type 'b

and 'a k =

| Cont : 'a * ('a, 'b) l -> 'b k

Task continuation of type 'b k: Upon completion, a task produces a value of some type 'a, and a list of subtasks from 'a to 'b.

and ('a, 'b) t = {

label : string option;

timeout : Mtime.span option;

task : 'a -> 'b k;

}

A single task.

val app : 'a 'b 'c. ('a, 'b) l -> ('b, 'c) l -> ('a, 'c) l

val of_list : ('a, 'b) t list -> ('b, 'c) l

val singleton : ('a, 'b) t -> ('a, 'c) l

val snoc : ('a, 'b) l -> ('c, 'd) t -> ('a, 'e) l

val then_ : ('a, 'b) t -> ('c, 'd) t -> ('e, 'f) l

type 'a st_in = {

x : 'a;

timeout : Mtime.span option;

parents : ctxs;

}

type 'a st_out = {

x : ('a, ctxs * task_failure) Stdlib.result;
(*
If the last task completed successfully with a value of x, then Ok x. Otherwise the context of the task that failed and the failure reason.
*)

passed : ctxs list;
(*
The tasks that completed successfully
*)

elapsed : Mtime.span;
(*
Elapsed time.
*)

}

val update_state_in : 'a -> Mtime.span -> 'b st_in -> 'c st_in

val merge_state_out : 'a st_out -> 'b st_out -> 'c st_out

val run_list : 'a 'b. 'a st_in -> ('a, 'b) l -> 'b st_out

Run a list of tasks. The result is as per st_out.

val run_task : 'a 'b. 'a st_in -> ('a, 'b) t -> 'b st_out

Run a tree of tasks. The result is as per st_out.

val run : ('a, 'b) t -> 'c -> 'b st_out

val task1 : ?timeout:Mtime.span -> ?label:string -> ('a -> 'b) -> ('c, 'd) t

val group : ?timeout:Mtime.span -> ?label:string -> ('a, 'b) l -> ('c, 'd) t

val with_ : ?timeout:Mtime.span -> ?label:string -> ('a, 'b) t -> ('c, 'd) t

type (_, _) unsnoc =

| Unsnoc : ('a, 'b) l * ('b, 'c) t -> ('a, 'c) unsnoc
(*
Unsnoc ts lst, is the list snoc ts lst, but split into ts and lst. The list ts contains all elements of the list except the last one, and lst is the last element.
*)

val unsnoc : 'a 'b 'c. ('a, 'b) t -> ('b, 'c) l -> ('a, 'c) unsnoc

unsnoc t1 ts splits the non-empty list cons t1 ts into its initial elements and final element, see unsnoc.

val then_map_k : 'a 'b. ('a -> 'b) -> 'a k -> 'b k

val then_map : 'a 'b 'c. ('b -> 'c) -> ('a, 'b) t -> ('a, 'c) t

Modify the final result of a task tree, without creating an additional task. Note: since f runs inside the final task, its side-effects should be minimal: any exceptions will be raised inside an existing task, and the existing task's timeout will include f. If that is undesirable, run f inside a new subtask with then_subtask.

val ignore : ('a, 'b) t -> ('a, unit) t

Ignore the result of a task (replace it with ()).

val first_map : ('a -> 'b) -> ('c, 'd) t -> ('e, 'f) t

val then_subtask : ('a, 'b) t -> ('c, 'd) t -> ('e, 'f) t

val accumulator : ('a -> 'b -> 'c) -> (unit, 'd) t -> ('e, 'f) t

val accumulating : ('a -> 'b -> 'c) -> 'd -> (unit, 'b) t list -> ('e, 'f) t

val collecting : (unit, 'a) t list -> ('b, 'c list) t

val pp_ctxs : Stdlib.Format.formatter -> ctxs -> unit

val pp_state_out : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'b st_out -> unit