Module Thread_counter.Counter

val lock_if : bool -> Stdlib.Mutex.t -> (unit, 'a, 'a) Common.Ctx_util.t

Note: we enforce that spawned threads don't raise uncaught exceptions, which in theory changes the semantics of threads. The value of being able to report stray exceptions outweighs the slim chance anyone would rely on being able to ignore exceptions in threads.

type 'a finished =

| Return of 'a

| Uncaught of Common.Util.exn_info

| Overflow of int

type 'a state =

| Running

| Finished of 'a finished

type 'a group = {

mutable state : 'a state;

finished : Stdlib.Condition.t;

mutable thread_count : int;

thread_limit : int option;

owner : t;

}

and g =

| G : 'a group -> g

and t = {

mut : Stdlib.Mutex.t;

groups : g ThreadH.t;

}

val finish : ?lock:bool -> 'a group -> 'a finished -> unit

val try_finish : 'a group -> 'a finished -> unit

val try_return : 'a group -> 'a -> unit

val spawn_thread : ?lock:bool -> ?group:'a group -> t -> ('b -> unit) -> 'b -> Thread.t

val create_counter : unit -> t

val create_group : ?thread_limit:int -> t -> 'a group

val get_thread_count : 'a group -> int

val join_group : leftover_thread_limit:int -> timeout:Mtime.span -> 'a group -> int

Wait for threads in a group to complete. Group must be finished first.

High-level group operations

type thread_group_err =

| ThreadLimitReached of int

| ThreadsLeftOver of {

left_over : int;

limit : int;

}

| ExceptionRaised of {

main : bool;

exn_info : Common.Util.exn_info;

}

val spawn_thread_group : ?thread_limit:int -> t -> ('a -> 'b) -> 'a -> ('b, Common.Util.exn_info) Stdlib.result group

Create a group that runs the given function, then sets the return value as the return value of the group.

val collect_thread_group : ?leftover_limit:(int * Mtime.span) -> 'a group -> 'a finished * int

Wait for a group to finish then join its threads (join_group).

val check_thread_group_result : ?leftover_limit:(int * Mtime.span) -> ('a, Common.Util.exn_info) Stdlib.result finished -> int -> ('a, thread_group_err list) Stdlib.result

Check the return value of collect_thread_group

val run_in_thread_group : ?thread_limit:int -> ?leftover_limit:(int * Mtime.span) -> t -> ('a -> 'b) -> 'a -> ('b, thread_group_err list) Stdlib.result

Combines spawn_thread_group, collect_thread_group, and check_thread_group_result.

val string_of_thread_group_err : thread_group_err -> string

val string_of_thread_group_errs : thread_group_err list -> string

exception ThreadGroupErrs of thread_group_err list

val thread_group_result_to_exn : ('a, thread_group_err list) Stdlib.result -> 'a