Less Power

Framework for sandboxed testing of OCaml code.

This framework consists of a number of libraries and executables for testing untrusted OCaml code. Designed for running programming exercises fully automatically and at scale.

Components

The following public components are included in the package:

Ast_check AST checker.

Test_lib Library of common functionality for writing tests.

Test_runner Library for creating a runner, which then drives the testing process.

Stdlib variants and overrides:

Stdlib_alerts Variants of Stdlib, but with signature items that can be restricted annotated with alerts.

Stdlib_components Adapted from Stdlib, with different components in different modules. In particular, a safe (pure) subset of OCaml features can be enabled by opening a subset of modules defined here.

Std_overrides Ready-to-use modules for replacing the default pervasives, using -open.

Internal Components

The following components are intended for use within framework code only:

Common Functionality common to multiple components.