Module Ast_check

AST checker.

Tool to enforce restrictions on what syntax elements are allowed in OCaml code. Restricts use of the following class features and imperative elements:

class declarations

class type declarations

class method calls

new class expressions

setting class and record fields

array literals

while loops

for loops

declaring records with mutable entries

external declarations

internal dune modules (Package__Module)

external declarations and internal dune modules are forbidden, as they can be used to circumvent restrictions in the Stdlib replacement. The sequence operator is not forbidden as there would be no point, a; b can be trivially replaced by let _ = a in b.

An executable is installed as lp-ast-check. All files passed as arguments, and all .ml files contained recursively in directories passed as arguments, are checked. Exits with code 0 if no violations are detected, 1 if violations are found, or 2 if an error occurs.

module Feature : sig ... end

type violation = {

location : Ppxlib.Location.t;
(*
Location of the violation.
*)

message : string option;
(*
Error message.
*)

feature : Feature.t;
(*
Which prohibited feature was used.
*)

}

A violation that occurred in the AST.

val ast_violations : ?prohibited:Feature.Set.t -> ?limit:int -> Ppxlib.structure -> violation list

Return a list of (up to limit) violations for this AST.

val file_violations : ?prohibited:Feature.Set.t -> ?limit:int -> string -> violation list

As per ast_violations, but first parse a file to an AST.

val path_violations : ?follow:FileUtil.action_link -> ?prohibited:Feature.Set.t -> ?limit:int -> ?check1:FileUtil.test_file -> ?check:FileUtil.test_file -> (violation list -> unit) -> string -> unit

As per file_violations, but scan an entire directory (recursively). If the path designates a regular file, check the file if check1 matches (default: always); if a directory is scanned, check each file if check matches (default: .ml ending). Pass each (possibly empty) list of violations to the callback; see the note in pp_violation.

The follow argument (default: Follow) only applies to directories; symlinks to files are always read.

val pp_violation : Stdlib.Format.formatter -> violation -> unit

Pretty-print a violation. Violations from a given file should be printed before the next file is parsed, otherwise no context from the source file will be output.

Breaks pretty-printing, this is a limitation of Location.print_report