Module Stdlib_variants

Variants of Stdlib with restricted or modified functionality.

module Stdlib_components : sig ... end

Adapted from Stdlib, with different components in different modules. In particular, a safe (pure) subset of OCaml features can be enabled by importing a subset of modules defined here.

module SafeStdlib : sig ... end

SAFE Everything safe from the standard library, including everything safe from sub-modules of Stdlib.