Less Power

Framework for sandboxed testing of OCaml code.

This framework consists of a number of libraries and executables for testing untrusted OCaml code. Designed for running programming exercises fully automatically and at scale.

Components

The following components are included in the package:

Ast_check AST checker.

Stdlib_variants Variants of Stdlib with restricted or modified functionality.

Test_lib Library of common functionality for writing tests.

Common Functionality common to multiple components.